move parent setting to own file
