+++ src/visitor.ts
@@ -1,5 +1,4 @@
 import * as ast from "./grammar/ast";
-import { ASTNode } from "./linting-rule";
 import { InputSource, Program } from "./semantics/program";
 import { TypeChecker } from "./semantics/type-checker";
 import { isKeyword } from "./util";
@@ -46,7 +45,7 @@ export class VisitorContext {
         this.scopeStack.shift();
     }
 
-    visit(node: ASTNode, callback: VisitorCallback) {
+    visit(node: ast.AnyAstNode, callback: VisitorCallback) {
         const result = callback(node, this);
 
         if (result !== VisitResult.StopVisiting) {
@@ -64,7 +63,7 @@ export enum VisitResult {
     Continue,
     StopVisiting,
 }
-export type VisitorCallback = (node: ASTNode, context: VisitorContext) => VisitResult | void;
+export type VisitorCallback = (node: ast.AnyAstNode, context: VisitorContext) => VisitResult | void;
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const visitors: Partial<Record<ast.SyntaxKind, (node: any, context: VisitorContext, cb: VisitorCallback) => void>> = {
@@ -120,6 +119,9 @@ const visitors: Partial<Record<ast.SyntaxKind, (node: any, context: VisitorConte
         cb: VisitorCallback
     ) => {
         context.pushScope(node);
+
+        context.visit(node.name, cb);
+
         for (const port of node.ports) {
             context.visit(port, cb);
         }
@@ -215,6 +217,9 @@ const visitors: Partial<Record<ast.SyntaxKind, (node: any, context: VisitorConte
         cb: VisitorCallback
     ) => {
         context.pushScope(node);
+
+        context.visit(node.name, cb);
+
         for (const type_or_event of node.body) {
             context.visit(type_or_event, cb);
         }
@@ -334,218 +339,7 @@ const visitors: Partial<Record<ast.SyntaxKind, (node: any, context: VisitorConte
     [ast.SyntaxKind.NumericLiteral]: stopVisiting,
 };
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const setParentVisitors: Partial<Record<ast.SyntaxKind, (node: any) => void>> = {
-    // Root node
-    [ast.SyntaxKind.File]: (node: ast.File) => {
-        for (const statement of node.statements) {
-            setParent(statement, node);
-        }
-    },
-
-    [ast.SyntaxKind.AssignmentStatement]: (node: ast.AssignmentStatement) => {
-        setParent(node.left, node);
-        setParent(node.right, node);
-    },
-    [ast.SyntaxKind.Behavior]: (node: ast.Behavior) => {
-        for (const statement of node.statements) {
-            setParent(statement, node);
-        }
-    },
-    [ast.SyntaxKind.BinaryExpression]: (node: ast.BinaryExpression) => {
-        setParent(node.left, node);
-        setParent(node.right, node);
-    },
-    [ast.SyntaxKind.Binding]: (node: ast.Binding) => {
-        setParent(node.left, node);
-        setParent(node.right, node);
-    },
-    [ast.SyntaxKind.BindingCompoundName]: (node: ast.BindingCompoundName) => {
-        setParent(node.compound, node);
-        setParent(node.name, node);
-    },
-    [ast.SyntaxKind.CallExpression]: (node: ast.CallExpression) => {
-        setParent(node.expression, node);
-        for (const expression of node.arguments.arguments) {
-            setParent(expression, node);
-        }
-    },
-    [ast.SyntaxKind.ComponentDefinition]: (node: ast.ComponentDefinition) => {
-        for (const port of node.ports) {
-            setParent(port, node);
-        }
-
-        if (node.body) {
-            setParent(node.body, node);
-        }
-    },
-    [ast.SyntaxKind.Compound]: (node: ast.Compound) => {
-        for (const statement of node.statements) {
-            setParent(statement, node);
-        }
-    },
-    [ast.SyntaxKind.CompoundName]: (node: ast.CompoundName) => {
-        if (node.compound) setParent(node.compound, node);
-        setParent(node.name, node);
-    },
-    [ast.SyntaxKind.DeferStatement]: (node: ast.DeferStatement) => {
-        if (node.arguments) {
-            for (const argument of node.arguments.arguments) {
-                setParent(argument, node);
-            }
-        }
-        setParent(node.statement, node);
-    },
-    [ast.SyntaxKind.EnumDefinition]: (node: ast.EnumDefinition) => {
-        setParent(node.name, node);
-        for (const member of node.members) {
-            setParent(member, node);
-        }
-    },
-    [ast.SyntaxKind.Event]: (node: ast.Event) => {
-        setParent(node.eventName, node);
-        setParent(node.type, node);
-
-        for (const param of node.parameters) {
-            setParent(param, node);
-            setParent(param.name, param);
-            setParent(param.type, param);
-        }
-    },
-    [ast.SyntaxKind.ExpressionStatement]: (node: ast.ExpressionStatement) => {
-        setParent(node.expression, node);
-    },
-    [ast.SyntaxKind.FunctionDefinition]: (node: ast.FunctionDefinition) => {
-        setParent(node.body, node);
-        setParent(node.returnType, node);
-        setParent(node.name, node);
-
-        for (const parameter of node.parameters) {
-            setParent(parameter, node);
-        }
-    },
-    [ast.SyntaxKind.FunctionParameter]: (node: ast.FunctionParameter) => {
-        setParent(node.type, node);
-        setParent(node.name, node);
-    },
-    [ast.SyntaxKind.GuardStatement]: (node: ast.GuardStatement) => {
-        if (node.condition) {
-            setParent(node.condition, node);
-        }
-
-        setParent(node.statement, node);
-    },
-    [ast.SyntaxKind.IfStatement]: (node: ast.IfStatement) => {
-        setParent(node.condition, node);
-        setParent(node.statement, node);
-
-        if (node.else) {
-            setParent(node.else, node);
-        }
-    },
-    [ast.SyntaxKind.Instance]: (node: ast.Instance) => {
-        setParent(node.name, node);
-        setParent(node.type, node);
-    },
-    [ast.SyntaxKind.InterfaceDefinition]: (node: ast.InterfaceDefinition) => {
-        for (const type_or_event of node.body) {
-            setParent(type_or_event, node);
-        }
-
-        if (node.behavior) {
-            setParent(node.behavior, node);
-            for (const statement of node.behavior.statements) {
-                setParent(statement, node.behavior);
-            }
-        }
-    },
-    [ast.SyntaxKind.InvariantStatement]: (node: ast.InvariantStatement) => {
-        setParent(node.expression, node);
-    },
-    [ast.SyntaxKind.Namespace]: (node: ast.Namespace) => {
-        while (node.name.kind !== ast.SyntaxKind.Identifier && node.name.compound) {
-            // De-sugar
-            const newNs = {
-                kind: ast.SyntaxKind.Namespace,
-                position: node.position,
-                name: { ...node.name.name, kind: ast.SyntaxKind.Identifier },
-                statements: node.statements,
-            } satisfies ast.Namespace;
-            node.name = node.name.compound;
-            node.statements = [newNs];
-        }
-        setParent(node.name, node);
-        for (const statement of node.statements) {
-            setParent(statement, node);
-        }
-    },
-    [ast.SyntaxKind.OnStatement]: (node: ast.OnStatement) => {
-        for (const trigger of node.triggers) {
-            if (trigger) setParent(trigger, node);
-        }
-
-        setParent(node.body, node);
-    },
-    [ast.SyntaxKind.OnParameter]: (node: ast.OnParameter) => {
-        setParent(node.name, node);
-
-        if (node.assignment) {
-            setParent(node.assignment, node);
-        }
-    },
-    [ast.SyntaxKind.OnTrigger]: (node: ast.OnTrigger) => {
-        if (isKeyword(node)) return;
-        setParent(node.name, node);
-
-        if (node.parameterList?.parameters) {
-            for (const parameter of node.parameterList.parameters) {
-                setParent(parameter, node);
-            }
-        }
-    },
-    [ast.SyntaxKind.ParenthesizedExpression]: (node: ast.ParenthesizedExpression) => {
-        setParent(node.expression, node);
-    },
-    [ast.SyntaxKind.Port]: (node: ast.Port) => {
-        setParent(node.name, node);
-        setParent(node.type, node);
-    },
-    [ast.SyntaxKind.Reply]: (node: ast.Reply) => {
-        if (node.port) setParent(node.port, node);
-        if (node.value) setParent(node.value, node);
-    },
-    [ast.SyntaxKind.ReturnStatement]: (node: ast.ReturnStatement) => {
-        if (node.returnValue) {
-            setParent(node.returnValue, node);
-        }
-    },
-    [ast.SyntaxKind.System]: (node: ast.System) => {
-        for (const instance_or_binding of node.instancesAndBindings) {
-            setParent(instance_or_binding, node);
-        }
-    },
-    [ast.SyntaxKind.TypeReference]: (node: ast.TypeReference) => {
-        setParent(node.typeName, node);
-    },
-    [ast.SyntaxKind.VariableDefinition]: (node: ast.VariableDefinition) => {
-        setParent(node.type, node);
-        if (node.initializer) {
-            setParent(node.initializer, node);
-        }
-    },
-    [ast.SyntaxKind.UnaryOperatorExpression]: (node: ast.UnaryOperatorExpression) => {
-        setParent(node.expression, node);
-    },
-};
-
 export function visitFile(file: ast.File, source: InputSource, callback: VisitorCallback, program: Program) {
     const context = new VisitorContext(source, program);
     context.visit(file, callback);
-}
-
-export const setParentVisitor: VisitorCallback = (node, context) =>
-    context.visit(node, n => setParentVisitors[n.kind]?.(n));
-
-function setParent(node: ASTNode, parent: ASTNode) {
-    node.parent = parent;
 }
